Check availability in op registration
